+++ Demo/Program.cs
@@ -1,16 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Demo
+﻿namespace Demo
 {
+	using System;
+
 	using ConsoleExtensions.Proxy;
 
-	class Program
+	public class Program
 	{
-		static void Main(string[] args)
+		public static void Header(IConsoleProxy console)
+		{
+			console.HR().WriteLine(" Fantastic 'ask your name' app").HR();
+		}
+
+		public static void Main(string[] args)
 		{
 			var console = ConsoleProxy.Instance();
 
@@ -18,13 +19,28 @@ namespace Demo
 
 			Styling(console);
 
-			console.ReadLine(out var t);
-
 			Header(console);
 
 			VanilaGreetAndAskForName();
 
 			GreetAndAskForName(console);
+			
+			console.ReadLine(out _);
+		}
+
+		public static void Styling(IConsoleProxy console)
+		{
+			console
+				.Style(StyleName.Ok)
+				.WriteLine("Mostly things are ok")
+				.Style(StyleName.Info)
+				.WriteLine("But sometimes you need to be informed")
+				.Style(StyleName.Warning)
+				.WriteLine("Or warned")
+				.Style(StyleName.Error)
+				.WriteLine("Or things can go really bad")
+				.ResetStyle()
+				.WriteLine("But mostly everything is fine.");
 		}
 
 		private static void GreetAndAskForName(IConsoleProxy console)
@@ -45,35 +61,6 @@ namespace Demo
 			var name = Console.ReadLine();
 			Console.WriteLine($"Welcome {name}");
 			Console.ReadLine();
-		}
-
-		public static void Header(IConsoleProxy console)
-		{
-			console.HR().WriteLine(" Fantastic 'ask your name' app").HR();
-		}
-
-		public static void Styling(IConsoleProxy console)
-		{
-			console
-				.Style(StyleName.Ok)
-				.WriteLine("Mostly things are ok")
-				.Style(StyleName.Info)
-				.WriteLine("But sometimes you need to be informed")
-				.Style(StyleName.Warning)
-				.WriteLine("Or warned")
-				.Style(StyleName.Error)
-				.WriteLine("Or things can go really bad")
-				.ResetStyle()
-				.WriteLine("But mostly everything is fine.");
-		}
-	}
-
-	public static class WriteExtensions
-	{
-		public static IConsoleProxy HR(this IConsoleProxy console)
-		{
-			console.Write(new string('-', console.WindowWidth));
-			return console;
 		}
 	}
 }
