--- TestProxy/Util.cs
@@ -1,10 +1,0 @@
-namespace ConsoleExtensions.Proxy.TestHelpers
-{
-  public static class Util
-  {
-    public static string[] LineSplit(TestProxy proxy)
-    {
-      return proxy.ToString().Replace("[key:", "\r[key:").Split('\r');
-    }
-  }
-}
